requirements.txt
streamlit
PyMuPDF
python-dotenv
langchain
langchain-community
langchain-google-genai
google-generativeai
faiss-cpu
